add test for list_content_types
