+++ unicore/distribute/tests/base.py
@@ -1,8 +1,11 @@
 import os
+import json
 
 from unittest import TestCase
 
 from elasticgit import EG
+from elasticgit.commands.avro import serialize
+from elasticgit.search import ESManager
 
 from unicore.distribute.utils import get_index_prefix
 
@@ -33,4 +36,29 @@ class DistributeTestCase(TestCase):
         while not workspace.index_ready():
             pass
 
+        return workspace
+
+    def add_schema(self, workspace, model_class):
+        schema_string = serialize(model_class)
+        schema = json.loads(schema_string)
+        workspace.sm.store_data(
+            os.path.join(
+                '_schemas',
+                '%(namespace)s.%(name)s.avsc' % schema),
+            schema_string, 'Writing the schema.')
+
+    def add_mapping(self, workspace, model_class):
+        im = ESManager(None, None, None)
+        mapping = im.get_mapping_type(model_class).get_mapping()
+        workspace.sm.store_data(
+            os.path.join(
+                '_mappings',
+                '%s.%s.json' % (model_class.__module__,
+                                model_class.__name__)),
+            json.dumps(mapping), 'Writing the mapping.')
+
+    def mk_model_workspace(self, model_class, *args, **kwargs):
+        workspace = self.mk_workspace(*args, **kwargs)
+        self.add_schema(workspace, model_class)
+        self.add_mapping(workspace, model_class)
         return workspace
